Refactored location + heading into an object for cleaner code
